apps/ai/pyproject.toml
[tool.poetry]
name = "personal ai"
version = "0.1.0"
description = ""
authors = ["SOUMITRO-SAHA <soumitrosahaofficial@gmail.com>"]
readme = "README.md"
packages = [{ include = "main.py", from = "." }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = { extras = ["standard"], version = "^0.115.2" }
langchain = "^0.3.3"
objectbox = "^4.0.0"
python-dotenv = "^1.0.1"
uvicorn = { extras = ["standard"], version = "^0.31.1" }


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
